tidy bird code into bird module
